% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_axes.R
\name{get_axes}
\alias{get_axes}
\title{Get major and minor axis}
\usage{
get_axes(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
sf object with exploded polygons
}
\description{
Get major axis (longest line between vertices) and minor axis (longest line inside a shape that is perpendicular to major axis) of polygon
}
\examples{
get_axes(vector_landscape, "class")
}
