% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_circum_diam.R
\name{get_circum_diam}
\alias{get_circum_diam}
\title{Diameter of smallest circumscribing circle(vector data)}
\usage{
get_circum_diam(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate diameter of smallest circumscribing circle
}
\details{
diameter of smallest circumscribing circle
}
\examples{
get_circum_diam(vector_patches, "class", "patch")
}
