% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eac_perim.R
\name{get_eac_perim}
\alias{get_eac_perim}
\title{Perimeter of equal-area circle(vector data)}
\usage{
get_eac_perim(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate perimeter of equal-area circle
}
\details{
perimeter of equal-area circle
}
\examples{
get_eac_perim(vector_patches, "class", "patch")
}
