% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hull_area.R
\name{get_hull_area}
\alias{get_hull_area}
\title{Area of the convex hull(vector data)}
\usage{
get_hull_area(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate area of the convex hull of the polygon
}
\details{
area of the convex hull of the polygon
}
\examples{
get_hull_area(vector_patches, "class", "patch")
}
