% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ibp}
\alias{get_ibp}
\title{Get boundary points of a shape}
\usage{
get_ibp(shape, n = 100)
}
\arguments{
\item{shape}{sf object}

\item{n}{number of points to generate}
}
\description{
Get boundary points of a shape
}
\keyword{internal}
