% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_igp}
\alias{get_igp}
\title{Get inner grid points for a shape}
\usage{
get_igp(shape, n = 1000)
}
\arguments{
\item{shape}{sf object}

\item{n}{number of points to generate}
}
\description{
Get inner grid points for a shape
}
\keyword{internal}
