% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landscape}
\alias{landscape}
\title{Example map (random cluster neutral landscape model).}
\format{A sf object.}
\source{
Simulated neutral landscape model with R. https://github.com/ropensci/NLMR/
}
\usage{
landscape
}
\description{
An example map to show landscapetools functionality
generated with the `nlm_randomcluster()` algorithm and
coerced to vector data (multipolygon).
}
\keyword{datasets}
