% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urban_atlas}
\alias{urban_atlas}
\title{Land use and land cover from Urban Atlas 2018}
\format{
A sf object.
}
\source{
Urban Atlas 2018
}
\usage{
urban_atlas
}
\description{
Land use and land cover from Urban Atlas 2018
}
\keyword{datasets}
