% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vector_patches}
\alias{vector_patches}
\title{vector_landscape object but with exploded multipolygons into polygons}
\format{
A sf object.
}
\source{
Simulated neutral landscape model with R.
}
\usage{
vector_patches
}
\description{
vector_landscape object but with exploded multipolygons into polygons
}
\keyword{datasets}
