% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_area_cv.R
\name{vm_c_area_cv}
\alias{vm_c_area_cv}
\title{The coefficient of variation of all patch areas at class level (vector data)}
\usage{
vm_c_area_cv(landscape, class_col)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}
}
\value{
the returned calculated coefficient of variation of areas of each class is in column "value",
this function returns also some important information such as level, class number and metric name.
}
\description{
This function allows you to calculate the coefficient of variation
of all patch areas belonging to one class in a categorical landscape in vector data format
}
\examples{
vm_c_area_cv(vector_landscape, "class")
}
