% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_area_mn.R
\name{vm_c_area_mn}
\alias{vm_c_area_mn}
\title{The mean value of all patch areas at class level(vector data)}
\usage{
vm_c_area_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated mean value of areas of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
}
\description{
This function allows you to calculate the mean value
of all patch areas belonging to one class in a categorical landscape in vector data format
}
\examples{
vm_c_area_mn(vector_landscape, "class")
}
