% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_ca.R
\name{vm_c_ca}
\alias{vm_c_ca}
\title{The total area of each class(vector data)}
\usage{
vm_c_ca(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated total class areas are in column "value",
and this function returns also some important information such as level, class number and metric name.
}
\description{
This function allows you to calculate the total area
of each class in a categorical landscape in vector data format
}
\examples{
vm_c_ca(vector_landscape, "class")
}
