% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_cai_cv.R
\name{vm_c_cai_cv}
\alias{vm_c_cai_cv}
\title{The coefficient of variation of ratio of the core area and the area at class level(vector data)}
\usage{
vm_c_cai_cv(landscape, class, core_distance)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{core_distance}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated coefficient of variation of ratio of the core area and the area of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the coefficient of variation
of the ratio of the core area and the area belonging to one class in a categorical landscape in vector data format
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_c_cai_cv(landscape, "landcover", core_distance = 1)
}
