% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_circle_sd.R
\name{vm_c_circle_sd}
\alias{vm_c_circle_sd}
\title{The standard deviation of Ratios between the patch area
and the smallest circumscribing circle of patches at class level(vector data)}
\usage{
vm_c_circle_sd(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated standard deviation of ratios of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the standard deviation of ratios of each class in a categorical landscape in vector data format
the ratio is the patch area relative to area of the smallest circumscribing circle of the patch
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_c_circle_sd(vector_landscape, "class")
}
