% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_coh_idx_mn.R
\name{vm_c_coh_idx_mn}
\alias{vm_c_coh_idx_mn}
\title{Cohesion Index(vector data)}
\usage{
vm_c_coh_idx_mn(landscape, class, n = 1000)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{n}{number of grid points to generate}
}
\value{
ratio of the average distance-squared among all points in an equalarea circle
and the average distance-squared among all points in the shape
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_coh_idx(vector_landscape, "class")
}
\description{
Calculate Cohesion Index
}
