% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_dcad.R
\name{vm_c_dcad}
\alias{vm_c_dcad}
\title{Disjunct core area density of each class(vector data)}
\usage{
vm_c_dcad(landscape, class_col, edge_depth)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{edge_depth}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated ratios are in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the number of disjunct core areas of each class
in relation to the landscape area in a categorical landscape in vector data format
}
\examples{
vm_c_dcad(vector_landscape, "class", edge_depth = 1)
}
