% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_division.R
\name{vm_c_division}
\alias{vm_c_division}
\title{Landscape division index of each class (vector data)}
\usage{
vm_c_division(landscape, class_col)
}
\arguments{
\item{landscape}{the input landscape}

\item{class_col}{the name of the class column}
}
\value{
A tibble with division index values (0-1)
}
\description{
Calculate the division index for each class, reflecting the probability
that two randomly selected points are not in the same patch.
}
\details{
Calculated as: \deqn{DIVISION = 1 - \sum (A_i / A_{landscape})^2}
Range: 0 to 1 (unitless). Higher values indicate more fragmentation.
}
\examples{
vm_c_division(vector_landscape, "class")
}
