% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_enn_cv.R
\name{vm_c_enn_cv}
\alias{vm_c_enn_cv}
\title{The coefficient of variation of Euclidean Nearest-Neighbor Distance of all patches at class level(vector data)}
\usage{
vm_c_enn_cv(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated coefficient of variation of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the coefficient of variation
of the Euclidean Nearest-Neighbor Distance among patches of one class in a categorical landscape in vector data format
Euclidean Nearest-Neighbor Distance means the distance from a patch edge to the nearest neighbouring patch belonging to the
same class.
}
\examples{
vm_c_enn_cv(vector_landscape, "class")
}
