% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_frac_cv.R
\name{vm_c_frac_cv}
\alias{vm_c_frac_cv}
\title{The coefficient of variation of fractal dimension index of all patches in each class(vector data)}
\usage{
vm_c_frac_cv(landscape, class_col)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}
}
\value{
the returned calculated coefficient of variation of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the coefficient of variation
of fractal dimension index of all patches belonging to class i in a categorical landscape in vector data format
The index is based on the patch perimeter and the patch area and describes the patch complexity
}
\examples{
vm_c_frac_cv(vector_landscape, "class")
}
