% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_full_idx_mn.R
\name{vm_c_full_idx_mn}
\alias{vm_c_full_idx_mn}
\title{Fullness Index(vector data)}
\usage{
vm_c_full_idx_mn(landscape, class, n = 10000)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{n}{number of local neighbourhoods to consider in calculating fullness}
}
\value{
ratio of the average fullness of small neighbourhoods (1% of area) in the shape and in its equal-area circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_fullness(vector_landscape, "class")
}
\description{
Calculate Fullness Index
}
