% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_fullness_mn.R
\name{vm_c_fullness_mn}
\alias{vm_c_fullness_mn}
\title{Fullness Index(vector data)}
\usage{
vm_c_fullness_mn(landscape, class_col, n = 1000)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{n}{number of local neighbourhoods to consider in calculating fullness}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Fullness Index
}
\details{
ratio between the average fullness of small neighbourhoods (1% of area) in the shape and in its equal-area circle
}
\examples{
vm_c_fullness_mn(vector_landscape, "class")
}
\references{
Angel, S., Parent, J., & Civco, D. L. (2010). Ten compactness properties of circles: Measuring shape in geography: Ten compactness properties of circles.
The Canadian Geographer / Le Géographe Canadien, 54(4), 441–461. https://doi.org/10.1111/j.1541-0064.2009.00304.x
}
