% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_girth_idx_mn.R
\name{vm_c_girth_idx_mn}
\alias{vm_c_girth_idx_mn}
\title{Girth Index(vector data)}
\usage{
vm_c_girth_idx_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio between radius of maximum inscribed circle and radius of equal-area circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_girth(vector_landscape, "class")
}
\description{
Calculate Girth Index
}
