% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_pafrac.R
\name{vm_c_pafrac}
\alias{vm_c_pafrac}
\title{Perimeter-Area Fractal Dimension(vector data)}
\usage{
vm_c_pafrac(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated slope is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to get the result of 2 divided by ß,
ß is the slope of the regression of the area against the perimeter in logarithm of each class in a categorical landscape in vector data format
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_c_pafrac(vector_landscape, "class")
}
