% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_pd.R
\name{vm_c_pd}
\alias{vm_c_pd}
\title{the patch density in each class(vector data)}
\usage{
vm_c_pd(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated density of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This metric is based on categorical landscape in vector data format. The density is the number of patches of each class
relative to the total landscape area. Then the number is standardised, so that the comparison among different landscape is possible.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_c_pd(vector_landscape, "class")
}
