% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_perarea_sd.R
\name{vm_c_perarea_sd}
\alias{vm_c_perarea_sd}
\title{The standard deviation of Perimeter-Area ratio index of all patches at class level(vector data)}
\usage{
vm_c_perarea_sd(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated standard deviation of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the standard deviation
of the ratios of all patches belonging to one class in a categorical landscape in vector data format
}
\examples{
vm_c_perarea_sd(vector_landscape, "class")
}
