% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_perim_idx_mn.R
\name{vm_c_perim_idx_mn}
\alias{vm_c_perim_idx_mn}
\title{Perimeter Index(vector data)}
\usage{
vm_c_perim_idx_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio between perimeter of equal-area circle and perimeter of polygon
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
st_p_area(vector_landscape, "class")
}
\description{
Calculate Perimeter Index
}
