% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_pland.R
\name{vm_c_pland}
\alias{vm_c_pland}
\title{Percentage of landscape of class(vector data)}
\usage{
vm_c_pland(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated percentage of each class is in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the percentage of each class in a categorical landscape in vector data format
That means each class area is standardised by the total landscape area, so the comparision among different landscape is possible.
}
\examples{
vm_c_pland(vector_landscape, "class")
}
