% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_split.R
\name{vm_c_split}
\alias{vm_c_split}
\title{Splitting index (vector data)}
\usage{
vm_c_split(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated indices are in column "value",
and this function returns also some important information such as level, class number and metric name.
Moreover, the "id" column, although it is just NA here at class level. we need it because the output struture of metrics
at class level should correspond to patch level one by one, and then it is more convinient to combine metric values at different levels and compare them.
}
\description{
This function allows you to calculate the relation between square of landscape area
and sum of square of all patch area of class i in a categorical landscape in vector data format
it is a aggregation metric.
}
\examples{
vm_c_split(vector_landscape, "class")
}
