% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_c_square_mn.R
\name{vm_c_square_mn}
\alias{vm_c_square_mn}
\title{Squareness(vector data)}
\usage{
vm_c_square_mn(landscape, class_col)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate squareness
}
\details{
ratio between perimeter of equal-area square of shape and perimeter of shape
}
\examples{
vm_c_square_mn(vector_landscape, "class")
}
\references{
Basaraner, M., & Cetinkaya, S. (2017). Performance of shape indices and classification schemes for characterising perceptual shape complexity of building footprints in GIS.
International Journal of Geographical Information Science, 31(10), 1952–1977. https://doi.org/10.1080/13658816.2017.1346257
}
