% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_core_mn.R
\name{vm_l_core_mn}
\alias{vm_l_core_mn}
\title{The mean value of all core areas in landscape(vector data)}
\usage{
vm_l_core_mn(landscape, edge_depth)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{edge_depth}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated mean value of core area of the whole landscape is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the mean value
of all core patch areas in a categorical landscape in vector data format
}
\examples{
vm_l_core_mn(vector_landscape, edge_depth= 1)
}
