% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_elong_idx_mn.R
\name{vm_l_elong_idx_mn}
\alias{vm_l_elong_idx_mn}
\title{Elongation(vector data)}
\usage{
vm_l_elong_idx_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio between major and minor axis length
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_elong_idx(vector_landscape, "class")
}
\description{
Calculate elongation of shape
}
