% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_elong_mn.R
\name{vm_l_elong_mn}
\alias{vm_l_elong_mn}
\title{Elongation(vector data)}
\usage{
vm_l_elong_mn(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate elongation of shape
}
\details{
ratio between major and minor axis length
}
\examples{
vm_l_elong_mn(vector_landscape)
}
\references{
Jiao, L., Liu, Y., & Li, H. (2012). Characterizing land-use classes in remote sensing imagery by shape metrics.
ISPRS Journal of Photogrammetry and Remote Sensing, 72, 46–55. https://doi.org/10.1016/j.isprsjprs.2012.05.012
}
