% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_exchange_mn.R
\name{vm_l_exchange_mn}
\alias{vm_l_exchange_mn}
\title{Exchange Index(vector data)}
\usage{
vm_l_exchange_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Exchange Index
}
\details{
share of the total area of the shape that is inside the equal-area circle about its centroid
}
\examples{
vm_l_exchange_mn(vector_landscape, "class")
}
\references{
Angel, S., Parent, J., & Civco, D. L. (2010). Ten compactness properties of circles: Measuring shape in geography: Ten compactness properties of circles.
The Canadian Geographer / Le Géographe Canadien, 54(4), 441–461. https://doi.org/10.1111/j.1541-0064.2009.00304.x
}
