% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_frac_mn.R
\name{vm_l_frac_mn}
\alias{vm_l_frac_mn}
\title{The mean value of the fractal dimension index of landscape(vector data)}
\usage{
vm_l_frac_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated mean value of the whole landscape is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the mean value
of fractal dimension index in a categorical landscape in vector data format
The index is based on the patch perimeter and the patch area and describes the patch complexity
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_frac_mn(vector_landscape, "class")
}
