% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_lpi.R
\name{vm_l_lpi}
\alias{vm_l_lpi}
\title{Largest patch index(vector data)}
\usage{
vm_l_lpi(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated index are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the maximal patch area in relative to
total landscape area in a categorical landscape in vector data format
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_lpi(vector_landscape, "class")
}
