% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_msidi.R
\name{vm_l_msidi}
\alias{vm_l_msidi}
\title{Modified Simpson's diversity index (vector data)}
\usage{
vm_l_msidi(landscape, class_col)
}
\arguments{
\item{landscape}{the input landscape image}

\item{class_col}{the name of the class column of the input landscape}
}
\value{
the returned calculated index is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the Modified Simpson's diversity index
in a categorical landscape in vector data format, Modified Simpson's diversity index is diversity index
}
\examples{
vm_l_msidi(vector_landscape, "class")
}
