% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_msiei.R
\name{vm_l_msiei}
\alias{vm_l_msiei}
\title{Modified Simpson's evenness index (vector data)}
\usage{
vm_l_msiei(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated index is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the Modified Simpson's evenness index
in a categorical landscape in vector data format, Modified Simpson's evenness index is diversity index
}
\examples{
vm_l_msiei(vector_landscape, "class")
}
