% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_pafrac.R
\name{vm_l_pafrac}
\alias{vm_l_pafrac}
\title{Perimeter-Area Fractal Dimension(vector data)}
\usage{
vm_l_pafrac(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated slope is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to get the result of 2 divided by ß,
ß is the slope of the regression of the area against the perimeter in logarithm of all patches in a categorical landscape in vector data format
}
\examples{
vm_l_pafrac(vector_landscape, "class")
}
