% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_pd.R
\name{vm_l_pd}
\alias{vm_l_pd}
\title{the patch density in the whole landscape(vector data)}
\usage{
vm_l_pd(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated density of each class is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This metric is based on categorical landscape in vector data format. The density is the number of patches of the whole landscape
relative to the total landscape area. Then the number is standardised, so that the comparison among different landscape is possible.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_pd(landscape, "landcover")
}
