% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_perarea_mn.R
\name{vm_l_perarea_mn}
\alias{vm_l_perarea_mn}
\title{Perimeter-Area ratio.}
\usage{
vm_l_perarea_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns the calculated ratio of all patches in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the ratio between the patch area and perimeter.
The ratio describes the patch complexity in a straightforward way.
}
\examples{
vm_l_perarea_mn(vector_landscape, "class")
}
