% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_ri_mn.R
\name{vm_l_ri_mn}
\alias{vm_l_ri_mn}
\title{Roughness index(vector data)}
\usage{
vm_l_ri_mn(landscape, class, n = 100)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{n}{number of boundary points to generate}
}
\value{
HERE WRITE DESCRIPTION OF METRIC
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_square(vector_landscape, "class")
}
\description{
Calculate Roughness index (RI)
}
