% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_rpr.R
\name{vm_l_rpr}
\alias{vm_l_rpr}
\title{Relative patch richness (vector data)}
\usage{
vm_l_rpr(landscape, class, class_max)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{class_max}{the maximal number of class in your input landscape image}
}
\value{
the returned calculated index is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the Relative patch richness
in a categorical landscape in vector data format, Relative patch richness is diversity index
}
\examples{
vm_l_rpr(vector_landscape, "class", 3)
}
