% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_shape_mn.R
\name{vm_l_shape_mn}
\alias{vm_l_shape_mn}
\title{The mean value of Shape index of landscape (vector data)}
\usage{
vm_l_shape_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated mean value of each class is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the mean value
of shape index of all patches in a categorical landscape in vector data format
shape index is the ratio between the actual perimeter of the patch and the hypothetical minimum perimeter of the patch.
The minimum perimeter equals the perimeter if the patch would be maximally compact. That means,
the perimeter of a circle with the same area of the patch.
}
\examples{
vm_l_shape_mn(vector_landscape, "class")
}
