% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_split.R
\name{vm_l_split}
\alias{vm_l_split}
\title{Splitting index (vector data)}
\usage{
vm_l_split(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated indices are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the relation between square of landscape area
and sum of square of all patch area in a categorical landscape in vector data format
it is a aggregation metric.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_split(landscape, "landcover")
}
