% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_ta.R
\name{vm_l_ta}
\alias{vm_l_ta}
\title{The total area of the whole landscape(vector data)}
\usage{
vm_l_ta(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the total area
of the whole landscape in a categorical landscape in vector data format
}
\examples{
vm_l_ta(vector_landscape)
}
