% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_tca.R
\name{vm_l_tca}
\alias{vm_l_tca}
\title{The total core area of the whole landscape(vector data)}
\usage{
vm_l_tca(landscape, edge_depth)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{edge_depth}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated total class core areas are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the total core area
in a categorical landscape in vector data format
}
\examples{
vm_l_tca(vector_landscape, edge_depth= 1)
}
