% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vm_metrics}
\alias{vm_metrics}
\title{Table with all metrics' names and descriptions}
\format{
data frame.
}
\usage{
vm_metrics
}
\description{
Table with all metrics' names and descriptions
}
\keyword{datasets}
