% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_circ.R
\name{vm_p_circ}
\alias{vm_p_circ}
\title{Circularity(vector data)}
\usage{
vm_p_circ(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Circularity
}
\details{
ratio between area of polygon and area of equal-perimeter circle
}
\examples{
vm_p_circ(vector_patches, "class", "patch")
}
\references{
Basaraner, M., & Cetinkaya, S. (2017). Performance of shape indices and classification schemes for characterising perceptual shape complexity of building footprints in GIS.
International Journal of Geographical Information Science, 31(10), 1952–1977. https://doi.org/10.1080/13658816.2017.1346257
}
