% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_circ_idx.R
\name{vm_p_circ_idx}
\alias{vm_p_circ_idx}
\title{Circularity(vector data)}
\usage{
vm_p_circ_idx(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio between area of polygon and area of equal-perimeter circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_circ_idx(vector_landscape, "class")
}
\description{
Circularity convexity
}
