% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_circle.R
\name{vm_p_circle}
\alias{vm_p_circle}
\title{Ratio between the patch area and the smallest circumscribing circle of the patch.}
\usage{
vm_p_circle(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns the calculated ratio of all patches in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the ratio between the patch area and the smallest circumscribing circle of the patch.
The diameter of the smallest circumscribing circle is the The distance between the two farthest points on the patch.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_circle(vector_landscape, "class")
}
