% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_circlep.R
\name{vm_p_circlep}
\alias{vm_p_circlep}
\title{Perimeter of equal-area circle(vector data)}
\usage{
vm_p_circlep(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
perimeter of equal-area circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
}
\description{
Calculate perimeter of equal-area circle
}
