% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_coh.R
\name{vm_p_coh}
\alias{vm_p_coh}
\title{Cohesion Index(vector data)}
\usage{
vm_p_coh(landscape, class_col = NULL, patch_col = NULL, n = 1000)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}

\item{n}{number of grid points to generate}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Cohesion Index
}
\details{
ratio of the average distance-squared among all points in an equal-area circle
and the average distance-squared among all points in the shape
}
\examples{
vm_p_coh(vector_patches, "class", "patch", 1000)
}
\references{
Angel, S., Parent, J., & Civco, D. L. (2010). Ten compactness properties of circles: Measuring shape in geography: Ten compactness properties of circles.
The Canadian Geographer / Le Géographe Canadien, 54(4), 441–461. https://doi.org/10.1111/j.1541-0064.2009.00304.x
}
