% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_conva.R
\name{vm_p_conva}
\alias{vm_p_conva}
\title{Area of the convex hull(vector data)}
\usage{
vm_p_conva(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
area of the convex hull of the polygon
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
}
\description{
Calculate area of the convex hull of the polygon
}
