% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_core.R
\name{vm_p_core}
\alias{vm_p_core}
\title{core area(vector data)}
\usage{
vm_p_core(landscape, class, core_distance)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{core_distance}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated core areas of all patches are in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the core area of all patches in square meters
Core area is defined as an area that within the patch and its edge is a fixed value from the boundary of the patch.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_core(vector_landscape, "class", core_distance = 0.8)
}
