% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_eac_perim.R
\name{vm_p_eac_perim}
\alias{vm_p_eac_perim}
\title{Perimeter of equal-area circle(vector data)}
\usage{
vm_p_eac_perim(landscape, class = NA, patch_id = NA)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{patch_id}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate perimeter of equal-area circle
}
\details{
perimeter of equal-area circle
}
\examples{
vm_p_eac_perim(vector_patches, "class", "patch")
}
