% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_eri.R
\name{vm_p_eri}
\alias{vm_p_eri}
\title{Equivalent rectangular index(vector data)}
\usage{
vm_p_eri(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio between perimeter of equal-area rectangle of shape and perimeter of shape
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_sq_idx(vector_landscape, "class")
}
\description{
Calculate Equivalent rectangular index (ERI)
}
