% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_frac.R
\name{vm_p_frac}
\alias{vm_p_frac}
\title{Fractal Dimension Index(vector data)}
\usage{
vm_p_frac(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate index fractal dimension index.
The index is based on the patch perimeter and the patch area and describes the patch complexity.
}
\details{
2 * log(perimeter) / log(area)
}
\examples{
vm_p_frac(vector_patches, "class", "patch")
}
