% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_full_idx.R
\name{vm_p_full_idx}
\alias{vm_p_full_idx}
\title{Fullness Index(vector data)}
\usage{
vm_p_full_idx(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio of the average fullness of small neighbourhoods (1% of area) in the shape and in its equal-area circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_full_idx(vector_landscape, "class")
}
\description{
Calculate Fullness Index
}
