% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_fullness.R
\name{vm_p_fullness}
\alias{vm_p_fullness}
\title{Fullness Index(vector data)}
\usage{
vm_p_fullness(landscape, class_col = NULL, patch_col = NULL, n = 10000)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}

\item{n}{number of local neighbourhoods to consider in calculating fullness}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Fullness Index
}
\details{
ratio between the average fullness of small neighbourhoods (1% of area) in the shape and in its equal-area circle
}
\examples{
vm_p_fullness(vector_patches, "class", "patch", 1000)
}
\references{
Angel, S., Parent, J., & Civco, D. L. (2010). Ten compactness properties of circles: Measuring shape in geography: Ten compactness properties of circles.
The Canadian Geographer / Le Géographe Canadien, 54(4), 441–461. https://doi.org/10.1111/j.1541-0064.2009.00304.x
}
