% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_girth.R
\name{vm_p_girth}
\alias{vm_p_girth}
\title{Girth Index(vector data)}
\usage{
vm_p_girth(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Girth Index
}
\details{
ratio between radius of maximum inscribed circle and radius of equal-area circle
}
\examples{
vm_p_girth(vector_patches, "class", "patch")
}
\references{
Angel, S., Parent, J., & Civco, D. L. (2010). Ten compactness properties of circles: Measuring shape in geography: Ten compactness properties of circles.
The Canadian Geographer / Le Géographe Canadien, 54(4), 441–461. https://doi.org/10.1111/j.1541-0064.2009.00304.x
}
