% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_hull_a.R
\name{vm_p_hull_a}
\alias{vm_p_hull_a}
\title{Area of the convex hull(vector data)}
\usage{
vm_p_hull_a(landscape, class = NA, patch_id = NA)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{patch_id}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate area of the convex hull of the polygon
}
\details{
area of the convex hull of the polygon
}
\examples{
vm_p_hull_a(vector_patches, "class", "patch")
}
