% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_hull_p.R
\name{vm_p_hull_p}
\alias{vm_p_hull_p}
\title{Perimeter of the convex hull(vector data)}
\usage{
vm_p_hull_p(landscape, class_col = NULL, patch_col = NULL)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate perimeter of the convex hull of the polygon
}
\details{
perimeter of the convex hull of the polygon
}
\examples{
vm_p_hull_p(vector_patches, "class", "patch")
}
