% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_ncore.R
\name{vm_p_ncore}
\alias{vm_p_ncore}
\title{the number of disjunct core area(vector data)}
\usage{
vm_p_ncore(landscape, class_col = NULL, patch_col = NULL, edge_depth)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class_col}{the name of the class column of the input landscape}

\item{patch_col}{the name of the id column of the input landscape}

\item{edge_depth}{the fixed distance to the edge of the patch}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the number of disjunct core areas of all patches
Core area is defined as an area that within the patch and its edge is a fixed value from the boundary of the patch.
Disjunct core area is defined as a new discrete area(patch), which is a sub-part of core area
}
\examples{
vm_p_core(vector_patches, "class", "patch", edge_depth = 0.8)
}
