% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_para_idx.R
\name{vm_p_para_idx}
\alias{vm_p_para_idx}
\title{Perimeter-Area ratio.}
\usage{
vm_p_para_idx(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns the calculated ratio of all patches in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the ratio between the patch area and perimeter.
The ratio describes the patch complexity in a straightforward way.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_para_idx(vector_landscape, "class")
}
