% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_perarea.R
\name{vm_p_perarea}
\alias{vm_p_perarea}
\title{Perimeter-Area ratio.}
\usage{
vm_p_perarea(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the ratio between the patch perimeter and area.
The ratio describes the patch complexity in a straightforward way.
}
\examples{
vm_p_perarea(vector_landscape, "class")
}
