% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_perim_idx.R
\name{vm_p_perim_idx}
\alias{vm_p_perim_idx}
\title{Perimeter Index(vector data)}
\usage{
vm_p_perim_idx(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Perimeter Index
}
\details{
ratio between perimeter of equal-area circle and perimeter of polygon
}
\examples{
vm_p_perim_idx(vector_landscape, "class")
}
\references{
Angel, S., Parent, J., & Civco, D. L. (2010). Ten compactness properties of circles: Measuring shape in geography: Ten compactness properties of circles.
The Canadian Geographer / Le Géographe Canadien, 54(4), 441–461. https://doi.org/10.1111/j.1541-0064.2009.00304.x
}
