% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_proxim_idx.R
\name{vm_p_proxim_idx}
\alias{vm_p_proxim_idx}
\title{Proximity Index(vector data)}
\usage{
vm_p_proxim_idx(landscape, class, n = 1000)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{n}{number of grid points to generate}
}
\value{
ratio between average distance from all points of equal-area circle to its center and average distance from all points of shape to its center
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_proxim_idx(vector_landscape, "class")
}
\description{
Calculate Proximity Index
}
