% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_range_idx.R
\name{vm_p_range_idx}
\alias{vm_p_range_idx}
\title{Range Index(vector data)}
\usage{
vm_p_range_idx(landscape, class, patch_id = NA)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
ratio between diameter of equal-area circle and diameter of smallest circumscribing circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
st_p_area(vector_landscape, "class")
}
\description{
Calculate Range Index
}
