% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{prepare_columns}
\alias{prepare_columns}
\title{Prepare class and patch ID columns}
\usage{
prepare_columns(landscape, class_col, patch_col)
}
\arguments{
\item{landscape}{sf object}

\item{class_col}{class column name}

\item{patch_col}{patch ID column name}
}
\description{
Prepare class and patch ID columns
}
\keyword{internal}
