% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectormetrics-package.R
\docType{package}
\name{vectormetrics-package}
\alias{vectormetrics}
\alias{vectormetrics-package}
\title{vectormetrics: Landscape Metrics for Categorical Map Patterns in Vector Data}
\description{
Package calculates landscape and shape metrics for categorical landscape patterns in a tidy workflow. 'vectormetrics' reimplements the most common landscapemetrics from 'FRAGSTATS' (\url{https://www.fragstats.org/}) and 'landscapemetrics' (\url{https://r-spatialecology.github.io/landscapemetrics/}) and adds many new metrics relating to shape of polygons, based on literature. This package supports Simple Feature objects as input arguments. It further provides utility functions to break multipolygon patches or calculate main axes of the shape.
}
\author{
\strong{Maintainer}: Tomasz Matuszek \email{tom.mateuszek@gmail.com}

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
  \item Marco Sciaini (\href{https://orcid.org/0000-0002-3042-5435}{ORCID})
  \item Maximillian H.K. Hesselbarth (\href{https://orcid.org/0000-0003-1125-9918}{ORCID})
  \item Yunyao Ma
}

}
\keyword{internal}
