% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_dcad.R
\name{vm_l_dcad}
\alias{vm_l_dcad}
\title{Disjunct core area density of the whole landscape(vector data)}
\usage{
vm_l_dcad(landscape, class, core_distance)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{core_distance}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated ratios are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the number of disjunct core areas
in relation to the landscape area in a categorical landscape in vector data format
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_dcad(landscape, "landcover", core_distance = 1)
}
