% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_dcore_mn.R
\name{vm_l_dcore_mn}
\alias{vm_l_dcore_mn}
\title{The mean value of number of core areas in landscape(vector data)}
\usage{
vm_l_dcore_mn(landscape, class, edge_depth)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{edge_depth}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated mean value of number of the whole landscape is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the mean value
of the total number of disjunct core areas in a categorical landscape in vector data format
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_dcore_mn(vector_landscape, "class", edge_depth= 1)
}
