% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_division.R
\name{vm_l_division}
\alias{vm_l_division}
\title{Landscape division index of the whole landscape(vector data)}
\usage{
vm_l_division(landscape)
}
\arguments{
\item{landscape}{the input landscape image}
}
\value{
the returned calculated index is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the Landscape division index
in a categorical landscape in vector data format, Landscape division index can somehow reflect
the probability that two randomly selected points are not located in the same patch
}
\examples{
vm_l_division(vector_landscape)
}
