% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_division.R
\name{vm_l_division}
\alias{vm_l_division}
\title{Landscape division index of the whole landscape(vector data)}
\usage{
vm_l_division(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated index is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the Landscape division index
in a categorical landscape in vector data format, Landscape division index can somehow reflect
the probability that two randomly selected points are not located in the same patch
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_division(landscape, "landcover")
}
