% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_ed.R
\name{vm_l_ed}
\alias{vm_l_ed}
\title{The Edge density of the whole landscape(vector data)}
\usage{
vm_l_ed(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated total length of perimeter is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the total length of all patches
in a categorical landscape in vector data format
}
\examples{
vm_l_ed(vector_landscape, "class")
}
