% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_frac_mn.R
\name{vm_l_frac_mn}
\alias{vm_l_frac_mn}
\title{fractal dimension index(vector data)}
\usage{
vm_l_frac_mn(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate index fractal dimension index.
The index is based on the patch perimeter and the patch area and describes the patch complexity.
}
\details{
2 * log(perimeter) / log(area)
}
\examples{
vm_l_frac_mn(vector_landscape, "class")
}
