% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_lpi.R
\name{vm_l_lpi}
\alias{vm_l_lpi}
\title{Largest patch index(vector data)}
\usage{
vm_l_lpi(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the maximal patch area in relative to
total landscape area in a categorical landscape in vector data format
}
\examples{
vm_l_lpi(vector_landscape)
}
