% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_lsi.R
\name{vm_l_lsi}
\alias{vm_l_lsi}
\title{Landscape shape index (vector data)}
\usage{
vm_l_lsi(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the returned calculated index are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
vm_l_lsi(vector_landscape)
}
\description{
This function allows you to calculate the ratio between the actual edge length of class i
and the hypothetical minimum edge length in a categorical landscape in vector data format.
The minimum edge length equals the edge length if class i would be maximally aggregated
}
