% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_lsi.R
\name{vm_l_lsi}
\alias{vm_l_lsi}
\title{Landscape shape index (vector data)}
\usage{
vm_l_lsi(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated index are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_lsi(vector_landscape, "class")
}
\description{
This function allows you to calculate the ratio between the actual edge length of class i
and the hypothetical minimum edge length in a categorical landscape in vector data format.
The minimum edge length equals the edge length if class i would be maximally aggregated
}
