% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_mesh.R
\name{vm_l_mesh}
\alias{vm_l_mesh}
\title{Effective Mesh Size (vector data)}
\usage{
vm_l_mesh(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated values are in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function helps to analyse the patch structure
the calculate process is, each patch is squared before the sums of them are calculated
and the sum is standardized by the total landscape area. it is a aggregation metric.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_mesh(vector_landscape, "class")
}
