% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_np.R
\name{vm_l_np}
\alias{vm_l_np}
\title{the number of patches in each class(vector data)}
\usage{
vm_l_np(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated number of patches of each class is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the number of patches of each class in a categorical landscape in vector data format
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_l_np(vector_landscape, "class")
}
