% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_pd.R
\name{vm_l_pd}
\alias{vm_l_pd}
\title{the patch density in the whole landscape(vector data)}
\usage{
vm_l_pd(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This metric is based on categorical landscape in vector data format. The density is the number of patches of the whole landscape
relative to the total landscape area. Then the number is standardised, so that the comparison among different landscape is possible.
}
\examples{
vm_l_pd(vector_landscape)
}
