% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_perarea_mn.R
\name{vm_l_perarea_mn}
\alias{vm_l_perarea_mn}
\title{The mean value of all patches Perimeter-Area ratio index at landscape level(vector data)}
\usage{
vm_l_perarea_mn(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the returned calculated mean value is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the mean value
of all patch ratios in a categorical landscape in vector data format
}
\examples{
vm_l_perarea_mn(vector_landscape)
}
