% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_l_te.R
\name{vm_l_te}
\alias{vm_l_te}
\title{Total (class) edge(vector data)}
\usage{
vm_l_te(landscape)
}
\arguments{
\item{landscape}{the input landscape image,}
}
\value{
the returned calculated length is in column "value",
and this function returns also some important information such as level and metric name,
Moreover, class number and the "id" column, although both are "NA" here in the landscape level
}
\description{
This function allows you to calculate the total length of edge of whole landscape in a categorical landscape in vector data format
}
\examples{
vm_l_te(vector_landscape)
}
