% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_cai.R
\name{vm_p_cai}
\alias{vm_p_cai}
\title{core area index(vector data)}
\usage{
vm_p_cai(landscape, class, core_distance)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{core_distance}{the fixed distance to the edge of the patch}
}
\value{
the returned calculated indices of all patches are in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the ratio of the core area and the area in square meters.
Core area is defined as an area that within the patch and its edge is a fixed value from the boundary of the patch.
The index describes the percentage of a patch that is core area.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_cai(landscape, "landcover", core_distance = 0.8)
}
