% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_circle.R
\name{vm_p_circle}
\alias{vm_p_circle}
\alias{vm_p_circle.sf}
\alias{vm_p_circle.SpatialPolygonsDataFrame}
\title{CIRCLE (patch level)}
\usage{
vm_p_circle(landscape, class)

\method{vm_p_circle}{sf}(landscape, class)

\method{vm_p_circle}{SpatialPolygonsDataFrame}(landscape, class)
}
\arguments{
\item{landscape}{*sf* MULTIPOLYGON or POLYGON feature}

\item{class}{Name of the class column of the input landscape}
}
\description{
Related Circumscribing Circle (Shape metric)
}
\details{
\deqn{CIRCLE = 1 - (\frac{a_{ij}} {a_{ij}^{circle}})}
where \eqn{a_{ij}} is the area in square meters and \eqn{a_{ij}^{circle}} the area of
the smallest circumscribing circle.

CIRCLE is a 'Shape metric'. The metric is the ratio between the patch area and the smallest
circumscribing circle of the patch. The diameter of the smallest circumscribing circle is
the 'diameter' of the patch connecting the opposing corner points of the polygon
that are the furthest away from each other. The metric characterises the compactness
of the patch and is comparable among patches with different area.

\subsection{Units}{None}
\subsection{Range}{0 <= CIRCLE < 1}
\subsection{Behaviour}{CIRCLE = 0 for a circular patch and approaches CIRCLE = 1 for
a linear patch.}
}
\examples{
-
vm_p_circle(vector_landscape, "class")

}
\references{
McGarigal, K., SA Cushman, and E Ene. 2012. FRAGSTATS v4: Spatial Pattern Analysis
Program for Categorical and Continuous Maps. Computer software program produced by
the authors at the University of Massachusetts, Amherst. Available at the following
web site: http://www.umass.edu/landeco/research/fragstats/fragstats.html

Baker, W. L., and Y. Cai. 1992. The r.le programs for multiscale analysis of
landscape structure using the GRASS geographical information system.
Landscape Ecology 7: 291-302.
}
