% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_circum.R
\name{vm_p_circum}
\alias{vm_p_circum}
\title{Diameter of smallest circumscribing circle(vector data)}
\usage{
vm_p_circum(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
diameter of equal-area circle
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
}
\description{
Calculate diameter of smallest circumscribing circle
}
