% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_eri.R
\name{vm_p_eri}
\alias{vm_p_eri}
\title{Equivalent rectangular index(vector data)}
\usage{
vm_p_eri(landscape, class = NA, patch_id = NA)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}

\item{patch_id}{the name of the id column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate Equivalent rectangular index (ERI)
}
\details{
ratio between perimeter of equal-area rectangle of shape and perimeter of shape
}
\examples{
vm_p_eri(vector_patches, "class", "patch")
}
\references{
Basaraner, M., & Cetinkaya, S. (2017). Performance of shape indices and classification schemes for characterising perceptual shape complexity of building footprints in GIS.
International Journal of Geographical Information Science, 31(10), 1952–1977. https://doi.org/10.1080/13658816.2017.1346257
}
