% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_perim.R
\name{vm_p_perim}
\alias{vm_p_perim}
\title{The perimeter of each patches(vector data)}
\usage{
vm_p_perim(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated perimeter of all patches is in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
st_p_area(vector_landscape, "class")
}
\description{
This function allows you to calculate the perimeter of each patches in a categorical landscape in vector data format
}
