% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_shape.R
\name{vm_p_shape}
\alias{vm_p_shape}
\title{Shape index(vector data)}
\usage{
vm_p_shape(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the returned calculated indices of all patches are in column "value",
and this function returns also some important information such as level, class, patch id and metric name.
}
\description{
This function allows you to calculate the shape index, which is
the ratio between the actual perimeter of the patch and the hypothetical minimum perimeter of the patch.
The minimum perimeter equals the perimeter if the patch would be maximally compact. That means,
the perimeter of a circle with the same area of the patch.
}
\examples{
## if the class name of input landscape is landcover,
## then write landcover in a double quotation marks as the second parameter.
vm_p_shape(vector_landscape, "class")
}
