% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vm_p_sph.R
\name{vm_p_sph}
\alias{vm_p_sph}
\title{Sphrecity(vector data)}
\usage{
vm_p_sph(landscape, class)
}
\arguments{
\item{landscape}{the input landscape image,}

\item{class}{the name of the class column of the input landscape}
}
\value{
the function returns tibble with the calculated values in column "value",
this function returns also some important information such as level, class, patch id and metric name.
}
\description{
Calculate sphercity
}
\details{
ratio between radius of maximum inscribed circle and minimum circumscribing circle
}
\examples{
vm_p_sph(vector_landscape, "class")
}
